UnloqAI-BE-dev/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy==2.0.23
pg8000==1.30.3
pydantic>=2.9.0,<3.0.0
anyio>=4.8.0,<5.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
mangum==0.17.0
pgvector==0.4.2
boto3==1.42.27
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
cryptography==41.0.7
bcrypt==4.1.2
psycopg2-binary==2.9.9
typing-extensions
google-genai
openai
